Modify IMerchPackRepository.cs

+++ src/OzonEdu.MerchandiseService.Domain/AggregationModels/MerchPackAggregate/IMerchPackRepository.cs
@@ -12,6 +12,6 @@ namespace OzonEdu.MerchandiseService.Domain.AggregationModels.MerchPackAggregate
         /// <param name="id">Идентификатор товарной позиции</param>
         /// <param name="cancellationToken">Токен для отмены операции. <see cref="CancellationToken"/></param>
         /// <returns>Товарная позиция</returns>
-        Task<MerchPack> FindByIdAsync(int id, CancellationToken cancellationToken = default);
+        Task<MerchPack> FindByIdAsync(long id, CancellationToken cancellationToken = default);
     }
 }
